Move non-integration tests to clients directory
